Some code lines add

+++ Assets/Done/Done_Scripts/Controller/Data/Database.cs
@@ -80,7 +80,7 @@ public class Database : MonoBehaviour {
         }
 
 
-        public float GetCellModelLevel ()
+        public int GetCellModelLevel ()
 		{
 			return this.cellModelLevel;
 		}
